requirements.txt
# Web Scraping
scrapy>=2.11.0
beautifulsoup4>=4.12.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Vector Store and Embeddings
faiss-cpu>=1.7.4
chromadb>=0.4.0
sentence-transformers>=2.2.0

# LLM and RAG
langchain>=0.1.0
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0

# Web Interface
streamlit>=1.29.0
gradio>=4.0.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0

# Development
jupyter>=1.0.0
pytest>=7.4.0
black>=23.10.0
flake8>=6.1.0 